requirements.txt
django-crispy-forms
python-dotenv
Pillow
psycopg2-binary